+++ src/main/java/calculator/ModOperator.java
@@ -2,7 +2,7 @@ package calculator;
 
 public class ModOperator implements Operator{
     @Override
-    public Double Operate(int firstNumber, int secondNumber) {
-        return (double)(firstNumber % secondNumber);
+    public Double Operate(double firstNumber, double secondNumber) {
+        return firstNumber % secondNumber;
     }
 }
