+++ src/main/java/calculator/Operator.java
@@ -1,5 +1,5 @@
 package calculator;
 
 public interface Operator {
-    Double Operate(int firstNumber,int secondNumber);
+    Double Operate(double firstNumber,double secondNumber);
 }
